Add XIM support
This allows use of composed characters in Xorg as well as use of external IMEs.
Restarting the IME is not currently supported (if an IME is detected on
startup, rterm will stop reading input if it exits).

External IME use is enabled using the "XMODIFIERS" environment variable,
handled by libX11, eg:
  XMODIFIERS=@im=SCIM